transport: convert u32 to u64 to get object

+++ drivers/block/cbd/cbd_transport.c
@@ -95,7 +95,7 @@ static inline struct cbd_##OBJ##_info						\
 										\
 	start += info->OBJ##_area_off;						\
 										\
-	return start + (info->OBJ_SIZE * id);					\
+	return start + ((u64)info->OBJ_SIZE * id);				\
 }										\
 										\
 struct cbd_##OBJ##_info								\
